requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via starlette
click==8.2.0
    # via uvicorn
colorama==0.4.6
    # via click
fastapi==0.115.12
    # via -r requirements.in
h11==0.16.0
    # via uvicorn
idna==3.10
    # via anyio
psycopg2-binary==2.9.10
    # via -r requirements.in
pydantic==2.11.4
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.9.1
    # via -r requirements.in
python-dotenv==1.1.0
    # via
    #   -r requirements.in
    #   pydantic-settings
sniffio==1.3.1
    # via anyio
starlette==0.46.2
    # via fastapi
typing-extensions==4.13.2
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.0
    # via
    #   pydantic
    #   pydantic-settings
uvicorn==0.34.2
    # via -r requirements.in
